Fixes en modificacion/baja
Si se intenta acceder al formulario de modificacion con una id
inexistente, se muestra el formulario para crear un cliente nuevo.
(antes causaba un error 500)

Si se intenta eliminar un usuario con id inexistente, el sistema ya no
tira un error 500 (pero tampoco muestra un error de ningun tipo).

+++ src/main/java/api/grupo/appservicios/service/impl/ClienteServiceImpl.java
@@ -46,7 +46,8 @@ public class ClienteServiceImpl implements ClienteService {
 
 	@Override
 	public void baja(long id) {
-		clienteDAO.deleteById(id);		
+		if (clienteDAO.existsById(id))
+			clienteDAO.deleteById(id);		
 	}
 
 }
